pyproject.toml
[project]
name = "wawa-kanban"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-fasthtml>=0.12.48",
]

[project.optional-dependencies]
test = ["pytest>=8.0"]
